* use wmo sub as triangle list
git-svn-id: http://localhost:88/svn/uwow/mpqnav/trunk@370 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ MPQNav/MPQ/WMO/Chunks/MOVI.cs
@@ -2,6 +2,6 @@ using System;
 
 namespace MPQNav.MPQ.WMO.Chunks {
 	internal class MOVI {
-		public short[] Indices { get; set; }
+		public int[] Indices { get; set; }
 	}
 }
