+++ MPQNav/MPQ/WMO/Chunks/Parsers/MOVIChunkParser.cs
@@ -12,7 +12,7 @@ namespace MPQNav.MPQ.WMO.Chunks.Parsers {
 		public override MOVI Parse() {
 			Reader.BaseStream.Position = AbsoluteStart;
 
-			var result = new List<short>();
+			var result = new List<int>();
 			while(Reader.BaseStream.Position < AbsoluteStart + Size) {
 				short one = Reader.ReadInt16();
 				short two = Reader.ReadInt16();
